Update Video Controllers

+++ src/controllers/video.controller.js
@@ -1,9 +1,18 @@
+import mongoose from "mongoose"
 import { Video } from "../models/video.model.js"
 import { ApiError } from "../utils/ApiError.js"
 import { ApiResponse } from "../utils/ApiResponse.js"
 import { asyncHandler } from "../utils/asyncHandler.js"
 import { deleteFromCloudinary, uploadOnCloudinary } from "../utils/cloudinary.js"
 
+
+const isVideoOwner = async (videoId,userId) => {
+    try {
+        return await Video.exists({_id: videoId,owner: userId})
+    } catch (error) {
+        throw error;
+    }
+}
 
 export const getAllVideos = asyncHandler(async (req, res) => {
     let { page = 1, limit = 10, query, sortBy = "views", sortType, userId } = req.query;
@@ -109,9 +118,135 @@ export const publishAVideo = asyncHandler(async (req, res) => {
 export const getVideoById = asyncHandler(async (req, res) => {
     const { videoId } = req.params
     //TODO: get video by id
-    const video = await Video.findById(videoId)
-    if (!video)
-        throw new ApiError(400, "Video not Found")
+    const video = await Video.aggregate([
+        {
+            $match:{
+                _id: new mongoose.Types.ObjectId(videoId)
+            }
+        },
+        {
+            $lookup:{
+                from: "users",
+                localField: "owner",
+                foreignField: "_id",
+                as: "user"
+            }
+        },
+        {
+            $unwind:"$user"
+        },
+        {
+            // Subscriber count and isSubscribed status
+            $lookup:{
+                from: "subscriptions",
+                let: {channelId:"$user._id" , userId: new mongoose.Types.ObjectId(req.user._id)},
+                pipeline:[
+                    {
+                        $match:{
+                            $expr:{
+                                $eq:['$channel','$$channelId']
+                            }
+                        }
+                    },
+                    {
+                        $group:{
+                            _id: null,
+                            subscriberCount: {$sum: 1},
+                            isSubscribed: {
+                                $max:{$eq:['$subscriber','$$userId']}
+                            }
+                        }
+                    },
+                ],
+                as: "subscribtionDetails"
+            }
+        },
+        {
+           $unwind:"$subscribtionDetails"
+        },
+        {
+            //Total Likes and Like status
+            $lookup:{
+                from: "likes",
+                let: {videoId: "$_id",userId:new mongoose.Types.ObjectId(req.user._id)},
+                pipeline:[
+                    {
+                        $match:{
+                            $expr:{
+                                $eq:['$video','$$videoId']
+                            }
+                            
+                        }
+                    },
+                    {
+                        $group:{
+                            _id: null,
+                            likeCount:{
+                                $sum : 1
+                            },
+                            likeStatus:{
+                                $max:{$eq:['$likedBy','$$userId']}
+                            }
+                        }
+                    }
+                ],
+                as: "likeDetails"
+
+            }
+        },
+        {
+            $unwind:{
+                path: '$likeDetails',
+                preserveNullAndEmptyArrays: true
+            }
+        },
+        {
+             //Total Likes and Like status
+             $lookup:{
+                from: "comments",
+                let: {videoId: "$_id"},
+                pipeline:[
+                    {
+                        $match:{
+                            $expr:{
+                                $eq:['$video','$$videoId']
+                            }
+                            
+                        }
+                    },
+                    {
+                        $count: 'commentCount'
+                    }
+                ],
+                as: 'commentDetails'
+             }
+        },
+        {
+            $unwind:'$commentDetails'
+        },
+        {
+            $project:{
+                videoFile: 1,
+                thumbnail: 1,
+                title: 1,
+                description: 1,
+                duration: 1,
+                views: 1,
+                createdAt: 1,
+                username: '$user.username',
+                avatar: '$user.avatar',
+                username: '$user.username',
+                isSubscribed: '$subscribtionDetails.isSubscribed',
+                channelSubs: '$subscribtionDetails.subscriberCount',
+                totalLikes: '$likeDetails.likeCount',
+                isLiked: '$likeDetails.likeStatus',
+                totalComments: '$commentDetails.commentCount'
+
+            }
+        }
+        
+        
+    ])
 
     return res.status(200).json(new ApiResponse(200, video, "Video Fetched Successfully."))
 })
@@ -119,11 +254,16 @@ export const getVideoById = asyncHandler(async (req, res) => {
 export const updateVideo = asyncHandler(async (req, res) => {
     const { videoId } = req.params
     //TODO: update video details like title, description, thumbnail
+    
     const { title, description } = req.body
     const thumbnailLocalPath = req.file
 
     if (!(title || thumbnailLocalPath || description))
         throw new ApiError(400, "Atleast one field required to update")
+
+    const isOwner = await isVideoOwner(videoId,req.user._id);
+    if(!isOwner)
+        throw new ApiError(400,"Not Authenticated")
 
     let thumbnail = "";
 
@@ -138,20 +278,25 @@ export const updateVideo = asyncHandler(async (req, res) => {
     if (thumbnail)
         await deleteFromCloudinary(prevVideoDetails.thumbnail)
 
-    prevVideoDetails.title = title ? title : prevVideoDetails.title;
-    prevVideoDetails.description = description ? description : prevVideoDetails.description;
-    prevVideoDetails.thumbnail = thumbnail ? thumbnail.url : prevVideoDetails.thumbnail
+   const updatedFields = {
+        title: title || prevVideoDetails.title,
+        description: description || prevVideoDetails.description,
+        thumbnail: thumbnail ? thumbnail.url : prevVideoDetails.thumbnail,
+    };
 
-    const newVideoDetails = await prevVideoDetails.save({ validateBeforeSave: false })
+    const newVideoDetails = await Video.findByIdAndUpdate(videoId, updatedFields, { new: true, runValidators: false });
 
     return res.status(200).json(new ApiResponse(200, newVideoDetails, "Changes Saved"))
-
 
 })
 
 export const deleteVideo = asyncHandler(async (req, res) => {
     const { videoId } = req.params
     //TODO: delete video
+    const isOwner = await isVideoOwner(videoId,req.user._id);
+    if(!isOwner)
+        throw new ApiError(400,"Not Authenticated")
+
     const deletedVideo = await Video.findByIdAndDelete(videoId)
     if (!deleteVideo)
         throw new ApiError(404, "Video not Found")
@@ -165,14 +310,19 @@ export const deleteVideo = asyncHandler(async (req, res) => {
 
 export const togglePublishStatus = asyncHandler(async (req, res) => {
     const { videoId } = req.params
-    const oldVideo = await Video.findById(videoId)
-    if (!oldVideo)
+
+    const isOwner = await isVideoOwner(videoId,req.user._id);
+    if(!isOwner)
+        throw new ApiError(400,"Not Authenticated")
+
+    const newVideo = await Video.findByIdAndUpdate(videoId,{
+        $set:{ isPublished: {$not: "$isPublished"}}
+        },
+        {new: true})
+
+    if (!newVideo)
         throw new ApiError(404, "Video Not Found")
 
-    oldVideo.isPublished = !oldVideo.isPublished
-
-    const newVideo = await oldVideo.save({ validateBeforeSave: false })
-
-
+  
     return res.status(200).json(new ApiResponse(200, newVideo, "Status Updated"))
 })
